Included bonus task functions
Added the addv function which first checks if the attempted entry is even an object, and then checks if it has the correct keys to be considered a pokemon. Added the searchName function which uses the array.filter function to check if any pokemon in our list has a name that matches the user's search. Both functions have been added to pokemonRepository.