Updated styles.css, index.html, and scripts.js
index.html
Added pokenumber heading to the modal so that I can display each pokemon's id.

styles.css
Added style rules so that the list now displays as a three-column list again. Moved media-queries to the bottom of the stylesheet as I realized some of them were unintentionally being overridden. Added style rules for new modal-poke-number content (the newly added modal segment for displaying pokemon IDs). Added new media query to account for the modal being cut off when viewing on a screen with 430px height or less (specifically, when viewing from a smartphone screen rotated to the side). Adjusted unknown style rule that was giving the list and list items 15px to padding-right, which would cause content to be off-center.

scripts.js
(Forgot to mention in last push) Updated button and modal text so that the names are now properly capitalized. Added item.id to the details retrieved so that I can display pokemon ID in the modal. Attempted to add id retrieval to display in the list as well, but this is currently non-functional; further assistance needed.